Fix waitpid and shell, la 🐐🐐 Co-authored-by: Uriel <urielsv@users.noreply.github.com>

+++ Kernel/include/scheduler.h
@@ -12,7 +12,7 @@ typedef struct scheduler_cdt* scheduler_adt;
 extern void asm_do_timer_tick();
 
 int32_t create_process(Function code, char **args, int argc, char *name, uint8_t priority, uint8_t unkillable);
-int kill_process(uint16_t pid);
+int kill_process(uint32_t pid);
 int block_process(uint64_t pid);
 int unblock_process(uint64_t pid);
 scheduler_adt init_scheduler();
